+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
-## v3.5.0
+## v3.5.1 - March 2nd, 2026
+### Added
+- Option to remove the "New Member" badge in servers [Miscellaneous]
+- Option to to remove the Gradient that shows on Text Input and Typing Indicator [Miscellaneous]
+    - Both scrolled and not scrolled
+
+### Fixed
+- Remove Nameplates removing Group DMs in DM list
+- Remove Profile Collections leaving a gap
+
+## v3.5.0 - February 18th, 2026
 ### Added
 - Option to remove user's custom statuses [Miscellaneous]
 - Option to remove user's status in profile view [Profile Customizations]